+++ database/migrations/2022_01_23_171431_create_estimates_table.php
@@ -36,9 +36,9 @@ class CreateEstimatesTable extends Migration
             $table->string('code');
             $table->string('full_number')->unique();
 
-            $table->float('price_ht')->default(0);
-            $table->float('price_total')->default(0);
-            $table->float('price_tva')->default(0);
+            $table->unsignedBigInteger('price_ht')->default(0);
+            $table->unsignedBigInteger('price_total')->default(0);
+            $table->unsignedBigInteger('price_tva')->default(0);
 
             $table->integer('status')->default(\App\Constants\Response::DEVIS_EN_ATTENTE);
